Add smart search for 'which-source', 'remove', 'release' and 'unrelease' commands
